+++ bika/lims/monkey/catalog.py
@@ -1,9 +1,19 @@
 # -*- coding: utf-8 -*-
 
-from bika.lims import api
 from plone.indexer.interfaces import IIndexableObject
 from Products.ZCatalog.ZCatalog import ZCatalog
 from zope.component import queryMultiAdapter
+
+from bika.lims import api
+from bika.lims.catalog import CATALOG_AUDITLOG
+
+
+def is_auditlog_enabled():
+    setup = api.get_setup()
+    # might happen during installation
+    if not setup:
+        return False
+    return setup.getEnableGlobalAuditlog()
 
 
 def catalog_object(self, object, uid=None, idxs=None,
@@ -12,6 +22,11 @@ def catalog_object(self, object, uid=None, idxs=None,
     # Never catalog temporary objects
     if api.is_temporary(object):
         return
+
+    # skip indexing auditlog catalog if disabled
+    if self.id == CATALOG_AUDITLOG:
+        if not is_auditlog_enabled():
+            return
 
     if idxs is None:
         idxs = []